Full python client works now!
